+++ .eslintrc.js
@@ -40,31 +40,6 @@ module.exports = {
         'ember/require-tagless-components': 'off',
       },
     },
-    // mixins
-    {
-      files: [
-        'addon/mixins/with-files.js',
-        'addon/services/file-queue.js',
-        'addon/queue.js',
-      ],
-      rules: {
-        'ember/no-mixins': 'off',
-        'ember/no-new-mixins': 'off',
-      },
-    },
-    // EmberObject, other misc deprecated patterns
-    {
-      files: [
-        'addon/system/data-transfer.js',
-        'addon/file.js',
-        'addon/queue.js',
-        'tests/unit/computed/sum-by-test.js',
-      ],
-      rules: {
-        'ember/no-classic-classes': 'off',
-        'ember/no-legacy-test-waiters': 'off',
-      },
-    },
     // node files
     {
       files: [
