--- addon/computed/sum-by.js
@@ -1,11 +1,0 @@
-import { computed, get } from '@ember/object';
-
-export default function (collectionKey, itemKey) {
-  return computed(`${collectionKey}.@each.${itemKey}`, function () {
-    let collection = get(this, collectionKey);
-
-    return collection.reduce(function (sum, item) {
-      return sum + get(item, itemKey);
-    }, 0);
-  });
-}
